+++ app/scripts/api/the100/drsThe100.service.js
@@ -2,8 +2,8 @@ import angular from 'angular';
 import _ from 'underscore';
 
 angular
-    .module('drsApp')
-    .factory('The100Service', The100Service);
+  .module('drsApp')
+  .factory('The100Service', The100Service);
 
 The100Service.$inject = ['$http', '$q'];
 
@@ -24,9 +24,10 @@ function The100Service($http, $q) {
     })
       .then($http)
       .then(function(response) {
-        const players = _.pluck(response.data, 'player');
-        // console.log("[drs] response: %o, player: %o", response, players);
-        return players;
+        const platform = response.data[0].platform;
+        const players = _.reject(_.pluck(response.data, 'player'), (item) => !angular.isDefined(item));
+        // console.log("[drs] response: %o, platform: %o, players: %o", response, platform, players);
+        return { platform: platform, players: players };
       })
       .catch(function(error) {
         return $q.reject(error);